# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: peter0928091516@gmail.com
    name: API Support
    url: http://www.flower-sharing.com/support
  description: A social media API for sharing flower photos and connecting with other
    flower enthusiasts.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Flower Sharing API
  version: "1.0"
paths:
...
  /post/{id}:
    delete:
      description: Delete a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete post
      tags:
      - posts
    get:
      description: Retrieve a single post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get post by ID
      tags:
      - posts
    put:
      consumes:
      - multipart/form-data
      description: Update a post by ID with select fields
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update (comma-separated)
        in: query
        name: select
        required: true
        type: string
      - description: Post title
        in: formData
        name: title
        type: string
      - description: Post content
        in: formData
        name: content
        type: string
      - description: Post image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - you are not the owner of this post
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update post by ID with select
      tags:
      - posts
  /post/{id}/dislike:
    delete:
      description: Remove a like from a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post disliked successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unlike a post
      tags:
      - posts
  /post/{id}/like:
    post:
      description: Add a like to a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post liked successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Like a post
      tags:
      - posts
  /post/{id}/likes:
    get:
      description: Retrieve the number of likes for a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post likes fetched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get post likes
      tags:
      - posts
  /post/all:
    get:
      description: Retrieve all posts
      produces:
      - application/json
      responses:
        "200":
          description: Posts fetched successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Posts not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all posts
      tags:
      - posts
  /post/pagination:
    get:
      description: Retrieve paginated list of posts
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Posts fetched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get posts with pagination
      tags:
      - posts
  /post/search:
    get:
      description: Search for posts by query string
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Posts searched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Search posts
      tags:
      - posts
  /post/user/{user_id}/all:
    get:
      description: Retrieve all posts by a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Posts fetched successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Posts not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all posts by user ID
      tags:
      - posts
  /post/user/{user_id}/liked:
    get:
      description: Retrieve the posts liked by a user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User liked posts fetched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid input
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get user liked posts
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"